Forge a protective MBR to make Windows accept a hybrid GPT
Windows requires a GPT partition table with a "pure" protective MBR
for tools like bcdedit to work properly (which in turn is required by
setup).
Unfortunately the BCM2837 boot ROM can't handle a "pure" GPT as needed
by Windows.
Work around this by writing a hybrid MBR on the SD card, and replacing
it with a fake protective MBR when UEFI or the OS tries to read it. With
the same hack applied to the Windows rpisdhc driver, Windows boots and
correctly recognizes the GPT.